requirements.txt
imgaug==0.4.0
numpy==1.21.3
opencv_python==4.5.4.58
pandas==1.3.4
Pillow==8.4.0
pytorch_lightning==1.4.9
torch==1.10.0
torchsummary==1.5.1
torchvision==0.11.1
tqdm==4.62.3
